Remove unnecessary file
